- make display md5sums twice

+++ qmerge.c
@@ -1,7 +1,7 @@
 /*
  * Copyright 2005-2006 Gentoo Foundation
  * Distributed under the terms of the GNU General Public License v2
- * $Header: /var/cvsroot/gentoo-projects/portage-utils/qmerge.c,v 1.49 2006/07/03 14:45:22 solar Exp $
+ * $Header: /var/cvsroot/gentoo-projects/portage-utils/qmerge.c,v 1.50 2006/07/04 14:01:32 solar Exp $
  *
  * Copyright 2005-2006 Ned Ludd        - <solar@gentoo.org>
  * Copyright 2005-2006 Mike Frysinger  - <vapier@gentoo.org>
@@ -39,8 +39,8 @@ static struct option const qmerge_long_opts[] = {
 };
 
 static const char *qmerge_opts_help[] = {
-	"Force download overwriting existing files",
-	"Force download overwriting existing files (skipping Packages)",
+	"Fetch package and newest Packages metadata",
+	"Fetch package (skipping Packages)",
 	"Search available packages",
 	"Install package",
 	"Uninstall package",
@@ -51,7 +51,7 @@ static const char *qmerge_opts_help[] = {
         COMMON_OPTS_HELP
 };
 
-static const char qmerge_rcsid[] = "$Id: qmerge.c,v 1.49 2006/07/03 14:45:22 solar Exp $";
+static const char qmerge_rcsid[] = "$Id: qmerge.c,v 1.50 2006/07/04 14:01:32 solar Exp $";
 #define qmerge_usage(ret) usage(ret, QMERGE_FLAGS, qmerge_long_opts, qmerge_opts_help, lookup_applet_idx("qmerge"))
 
 char search_pkgs = 0;
@@ -969,9 +969,11 @@ int pkg_verify_checksums(char *fname, struct pkg_t *pkg, depend_atom *atom, int 
 	if (pkg->MD5[0]) {
 		hash = (char*) hash_file(fname, HASH_MD5);
 		if (strcmp(hash, pkg->MD5) == 0) {
-			qprintf("MD5:  [%sOK%s] %s %s/%s\n", GREEN, NORM, hash, atom->CATEGORY, pkg->PF);
+			if (display)
+				qprintf("MD5:  [%sOK%s] %s %s/%s\n", GREEN, NORM, hash, atom->CATEGORY, pkg->PF);
 		} else {
-			warn("MD5:  [%sER%s] (%s) != (%s) %s/%s", RED, NORM, hash, pkg->MD5, atom->CATEGORY, pkg->PF);
+			if (display)
+				warn("MD5:  [%sER%s] (%s) != (%s) %s/%s", RED, NORM, hash, pkg->MD5, atom->CATEGORY, pkg->PF);
 			ret++;
 		}
 	}
@@ -1029,12 +1031,9 @@ void pkg_fetch(int argc, char **argv, struct pkg_t *pkg) {
 		/* check to see if file exists and it's checksum matches */
 		snprintf(buf, sizeof(buf), "%s/%s.tbz2", pkgdir, pkg->PF);
 		unlink_empty(buf);
-		if (force_download) {
-			if (access(buf, R_OK) == 0)
-				if ((pkg->SHA1[0]) || (pkg->MD5[0]))
-					if ((pkg_verify_checksums(buf, pkg, atom, 0, 0)) != 0)
-						unlink(buf);
-		}
+		if ((force_download) && (access(buf, R_OK) == 0) && ((pkg->SHA1[0]) || (pkg->MD5[0])))
+			if ((pkg_verify_checksums(buf, pkg, atom, 0, 0)) != 0)
+				unlink(buf);
 
 		if (access(buf, R_OK) == 0) {
 			if ((!pkg->SHA1[0]) && (!pkg->MD5[0])) {
